ac: use bigger storage for ac_arg::arg_index / ac_shader_args::arg_count
AC_MAX_ARGS is now 384 so uint8_t isn't enough.

Fixes: 6f130342658 ("ac/llvm: prepare for passing VS->TCS IO via VGPRs")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/7750>

+++ src/amd/common/ac_shader_args.h
@@ -47,7 +47,7 @@ enum ac_arg_type
 };
 
 struct ac_arg {
-   uint8_t arg_index;
+   uint16_t arg_index;
    bool used;
 };
 
@@ -63,7 +63,7 @@ struct ac_shader_args {
       bool skip;
    } args[AC_MAX_ARGS];
 
-   uint8_t arg_count;
+   uint16_t arg_count;
    uint8_t sgpr_count;
    uint8_t num_sgprs_used;
    uint8_t num_vgprs_used;
